コールバック処理がなんかおかしかったので修正
